Big Changes to Explosives
Anything attached to the explosives now can get attached to a function I've made that also makes them disappear. It's not the most optimized way, I've realized upon finishing it there is already a more optimal way, but it works. Had to change the UProperty for StaticMesh however to make it work, from VisibleDefaultsOnly to EditDefaultsOnly, BlueprintReadWrite.